pyproject.toml
[project]
name = "personal-assistant-desktop-server"
version = "0.1.0"
description = ""
authors = [
    {name = "joegutman.dev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "quart (>=0.20.0,<0.21.0)",
    "hypercorn (>=0.17.3,<0.18.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "sqlalchemy-utils (>=0.41.2,<0.42.0)",
    "marshmallow (>=4.0.0,<5.0.0)",
    "marshmallow-sqlalchemy (>=1.4.2,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "cryptography (>=45.0.5,<46.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "transformers (>=4.53.1,<5.0.0)",
    "sentencepiece (>=0.2.0,<0.3.0)",
    "huggingface-hub (>=0.33.2,<0.34.0)",
    "accelerate (>=1.8.1,<2.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "faster-whisper (>=1.1.1,<2.0.0)",
    "sentence-transformers (>=5.0.0,<6.0.0)",
    "soundfile (>=0.13.1,<0.14.0)",
    "scipy (>=1.16.0,<2.0.0)",
    "numpy (>=2.3.1,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "bitsandbytes (>=0.46.1,<0.47.0)",
    "piper-tts (>=1.3.0,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
]

[tool.poetry]
package-mode = true
packages = [{include = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
